refactor(value): split Number and Mapping into submodules (#105)
value.rs 4637 -> 2857 lines; Number and Mapping/MappingAny moved to
value/ submodules. Pure mechanical move, no API change (semver-checks
clean). Full suite + Miri green.
